extract basetrack
